+++ R/hash.R
@@ -6,7 +6,7 @@
 #' @param x Character vector. If not a character vector, then
 #' [as.character()] is used to try to coerce it into one. `NA` entries
 #' will have an `NA` hash.
-#' @return `hash_sha256()` returns aharacter vector of hexadecimal
+#' @return `hash_sha256()` returns a character vector of hexadecimal
 #' SHA-256 hashes.
 
 #' @family hash functions
@@ -69,6 +69,76 @@ hash_file_sha256 <- function(paths) {
   .Call(clic_sha256_file, paths)
 }
 
+#' SHA-1 hash
+#'
+#' Calculate the SHA-1 hash of each element of a character vector.
+#'
+#' @param x Character vector. If not a character vector, then
+#' [as.character()] is used to try to coerce it into one. `NA` entries
+#' will have an `NA` hash.
+#' @return `hash_sha1()` returns a character vector of hexadecimal
+#' SHA-1 hashes.
+
+#' @family hash functions
+#'
+#' @export
+#' @examples
+#' hash_sha1(c("foo", NA, "bar", ""))
+
+hash_sha1 <- function(x) {
+  if (!is.character(x)) x <- as.character(x)
+  na <- is.na(x)
+  x[na] <- NA_character_
+  x[!na] <- .Call(clic_sha1, x[!na])
+  x
+}
+
+#' @export
+#' @rdname hash_sha1
+#' @details `hash_raw_sha1()` calculates the SHA-1 hash of the bytes
+#' of a raw vector.
+#'
+#' @return `hash_raw_sha1()` returns a character scalar.
+
+hash_raw_sha1 <- function(x) {
+  stopifnot(is.raw(x))
+  .Call(clic_sha1_raw, x)
+}
+
+#' @export
+#' @rdname hash_sha1
+#' @param serialize_version Workspace format version to use, see
+#' [base::serialize()].
+#' @details `hash_obj_sha1()` calculates the SHA-1 hash of an R
+#' object. The object is serialized into a binary vector first.
+#'
+#' @return `hash_obj_sha1()` returns a character scalar.
+
+hash_obj_sha1 <- function(x, serialize_version = 2) {
+  sr <- serialize(x, NULL, version = serialize_version)[-(1:14)]
+  hash_raw_sha1(sr)
+}
+
+#' @export
+#' @rdname hash_sha1
+#' @param paths Character vector of file names.
+#' @details `hash_file_sha1()` calculates the SHA-1 hash of one or
+#' more files.
+#'
+#' @return `hash_file_sha1()` returns a character vector of SHA-1
+#' hashes.
+
+hash_file_sha1 <- function(paths) {
+  if (!is.character(paths)) paths <- as.character(paths)
+  paths <- normalizePath(paths, mustWork = FALSE)
+  if (is_windows()) {
+    paths <- enc2utf8(paths)
+  } else {
+    paths <- enc2native(paths)
+  }
+  .Call(clic_sha1_file, paths)
+}
+
 #' MD5 hash
 #'
 #' Calculate the MD5 hash of each element of a character vector.
@@ -121,6 +191,23 @@ hash_obj_md5 <- function(x, serialize_version = 2) {
   hash_raw_md5(sr)
 }
 
+#' @export
+#' @rdname hash_md5
+#' @param paths Character vector of file names.
+#' @details `hash_file_md5()` calculates the MD5 hash of one or more
+#' files.
+
+hash_file_md5 <- function(paths) {
+  if (!is.character(paths)) paths <- as.character(paths)
+  paths <- normalizePath(paths, mustWork = FALSE)
+  if (is_windows()) {
+    paths <- enc2utf8(paths)
+  } else {
+    paths <- enc2native(paths)
+  }
+  .Call(clic_md5_file, paths)
+}
+
 #' Emoji hash
 #'
 #' @details
@@ -149,7 +236,7 @@ hash_obj_md5 <- function(x, serialize_version = 2) {
 #'
 #' @param x Character vector. `NA` entries will have an `NA` hash.
 #' @param size Number of emojis to use in a hash. Currently it has to
-#'   be between 1 and 4.
+#'   be from 1 through 4.
 #' @return `hash_emoji()` returns a data frame with columns
 #'   * `hash`: the emoji hash, a string of the requested size.
 #'   * `emojis`: list column with the emoji characters in character
@@ -291,7 +378,7 @@ hash_obj_emoji <- function(x, size = 3, serialize_version = 2) {
 #' }
 #' ```
 #'
-#' `hash_animals()` uses `r length(gfycat_animals)` animal names and
+#' `hash_animal()` uses `r length(gfycat_animals)` animal names and
 #' `r length(gfycat_adjectives)` different adjectives. The number of
 #' different hashes you can get for different values of `n_adj`:
 #'
@@ -310,7 +397,7 @@ hash_obj_emoji <- function(x, size = 3, serialize_version = 2) {
 #' from <https://gfycat.com>.
 #'
 #' @param x Character vector. `NA` entries will have an `NA` hash.
-#' @param n_adj Number of adjectives to use. It must be between 0 and 3.
+#' @param n_adj Number of adjectives to use. It must be from 0 through 3.
 #' @return A data frame with columns
 #'   * `hash`: the hash value, a string.
 #'   * `words`: list column with the adjectives and the animal name in a
@@ -381,7 +468,7 @@ hash_animal1_transform <- function(md5, n_adj) {
 
 #' @export
 #' @rdname hash_animal
-#' @details `hash_raw_anima()` calculates the adjective-animal hash of
+#' @details `hash_raw_animal()` calculates the adjective-animal hash of
 #' the bytes of a raw vector.
 #'
 #' @return `hash_raw_animal()` and `hash_obj_animal()` return a list