+++ NEWS.md
@@ -1,5 +1,29 @@
 # cli (development version)
 
+# cli 3.6.1
+
+* ANSI hyperlinks are now turned off on the RStudio render plane (#581).
+
+# cli 3.6.0
+
+* The progressr progress handler now reports progress correctly
+  (@HenrikBengtsson, #558).
+
+* New `hash_*sha1()` functions to calculate the SHA-1 hash of strings,
+  objects, files.
+
+* cli now shows progress bars after one second by default, if they
+  are less than half way at the point. (Or after two seconds,
+  unconditionally, as before.) See the the `cli.progress_show_after`
+  option in `?cli-config` for details (#542).
+
+* `format_inline()` now has a new argument `keep_whitespace`, and it keeps
+  whitespace, including newline and form feed characters by default.
+
+# cli 3.5.0
+
+* New `keypress()` function to read a single key press from a terminal.
+
 * New function `pretty_print_code()` to print function objects with syntax
   highlighting at the R console.
 
@@ -10,6 +34,9 @@
 
 * `ansi_strtrim()` now handles some edge cases better, when `ellipsis` has
   length zero, and when it is wider than `width`.
+
+* New `hash_file_md5()` function to calculate the MD5 hash of one or more
+  files.
 
 # cli 3.4.1